requirements.txt
rdkit-pypi==2022.9.4
numpy==1.22
gensim==3.8.3
mol2vec
scikit-learn==0.24.0
pandas==1.2.0
tqdm
prefetch-generator
torch==1.10.0
